+++ lib/city-manager.js
@@ -2,8 +2,8 @@
 "use strict";
 const path = require('path');
 const fs = require('fs');
-const Savegame = require('./savegame');
-const Index = require('./index');
+const Savegame = require('./savegame.js');
+const FileIndex = require('./file-index.js');
 const SC4 = path.resolve(process.env.HOMEPATH, 'documents/SimCity 4');
 const regions = path.join(SC4, 'regions');
 const plugins = path.join(SC4, 'plugins');
