+++ lib/file-index.js
@@ -1,4 +1,4 @@
-// # index.js
+// # file-index.js
 "use strict";
 const fs = require('fs');
 const path = require('path');
@@ -16,14 +16,14 @@ if (!fs.promises) {
 	};
 }
 
-// # Index
-// The index is a data structure that scans a list of dbpf files and builds up 
-// an index of all files in it by their TGI's. This should mimmick how the 
-// game scans the plugins folder as well. We obivously cannot keep everything 
-// in memory so we'll keep pointers to where we can find each file **on the 
-// disk**. Note: we should make use of node's async nature here so that we can 
-// read in as much files as possible in parallel!
-class Index {
+// # FileIndex
+// The file index is a data structure that scans a list of dbpf files and 
+// builds up an index of all files in it by their TGI's. This should mimmick 
+// how the game scans the plugins folder as well. We obivously cannot keep 
+// everything in memory so we'll keep pointers to where we can find each file 
+// **on the disk**. Note: we should make use of node's async nature here so 
+// that we can read in as much files as possible in parallel!
+class FileIndex {
 
 	// ## constructor(opts)
 	constructor(opts) {
@@ -127,14 +127,14 @@ class Index {
 	}
 
 }
-module.exports = Index;
+module.exports = FileIndex;
 
 // Object that we'll re-use to query so that we don't have to recreate it all 
 // the time.
 const query = {
-	"type": 0,
-	"group": 0,
-	"instance": 0
+	type: 0,
+	group: 0,
+	instance: 0
 };
 
 // # Record
@@ -171,7 +171,8 @@ class Record extends Entry {
 		if (this.file) return this.file;
 		if (this.raw) return this.raw;
 
-		// Read from the file, but at the correct offset (and in a syncronous way).
+		// Read from the file, but at the correct offset (and in a synchronous 
+		// way).
 		let file = String(this.source);
 		let fd = fs.openSync(file, 'r');
 		let buff = Buffer.allocUnsafe(this.compressedSize);
